Update patch set 1
Patch Set 1:

> Patch Set 1:
 > 
 > Why not update the original gerrit instead of creating anew ?
 > 
 > Also, why drop the [m/n] tags ?  They're useful to inform people that they have all the patches (in conjunction with topics).

The change-id is so that we can trace this change back to previous releases.

I find the [m/n] tags somewhat extraneous considering we have topics, and these topics can be easily found by searching the change-id of a change. However, if you find it useful, I'll add it back in a subsequent patchset.

Patch-set: 1