Update patch set 21
Patch Set 21: Code-Review-1

Sorry to be that guy, but I'm not sure the average person would find this all that useful. It causes more clutter on the status bar and adds more marginally useful toggles, circa cm7. Also previous discussions on this feature resulted in not shipping it (if that matters at all).

Patch-set: 21