Update patch set 1
Patch Set 1:

If the majority wants it disabled, we disable it.
If the majority wants it enabled, we enable it.

So maybe there should be a vote, but personally I have a feeling that the majority prefers it enabled.

And also, our implementation is a gazillion times better than the OEM's.

Patch-set: 1
Reviewer: Arne Coucheron <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0